# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Picroom API
  version: 1.0.0
  description: |
    Picroom — self-hosted image hosting service.

    Two API surfaces:
    1. REST/JSON under `/api/v1/` — first-class management API.
    2. AWS S3-compatible under `/s3/` — drop-in replacement for object storage
       clients (PicGo, rclone, AWS CLI).

    All endpoints require authentication except where noted.
  license:
    name: MIT
    identifier: MIT
  contact:
    name: Picroom Maintainers
    url: https://github.com/picroom/picroom
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://{host}
    description: Self-hosted production
    variables:
      host:
        default: picroom.example.com
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: Password login
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Session cookie (HttpOnly, SameSite=Lax)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/auth/oidc/{provider}/login:
    get:
      tags: [auth]
      summary: Begin OIDC login flow
      operationId: oidcLoginStart
      security: []
      parameters:
        - in: path
          name: provider
          required: true
          schema:
            type: string
          description: OIDC provider identifier
      responses:
        '302':
          description: Redirect to OIDC provider
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/auth/oidc/{provider}/callback:
    get:
      tags: [auth]
      summary: OIDC callback
      operationId: oidcCallback
      security: []
      parameters:
        - in: path
          name: provider
          required: true
          schema:
            type: string
        - in: query
          name: code
          required: true
          schema:
            type: string
        - in: query
          name: state
          required: true
          schema:
            type: string
      responses:
        '302':
          description: Redirect to application with session cookie
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: picroom_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    sigv4:
      type: apiKey
      in: header
      name: Authorization
      description: AWS Signature Version 4
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    RateLimited:
      description: Too many requests
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
          examples: [quota_exceeded, validation_failed]
        message:
          type: string
          description: Human-readable message
        request_id:
          type: string
          format: uuid
        details:
          type: object
          additionalProperties: true
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        totp:
          type: string
          description: Optional TOTP code
    User:
      type: object
      required: [id, email, name, role, created_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [viewer, uploader, manager, admin]
        avatar_url:
          type: string
          format: uri
          nullable: true
        created_at:
          type: string
          format: date-time
...
security:
  - cookieAuth: []
  - bearerAuth: []